requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
cryptography==37.0.4
h11==0.13.0
idna==3.3
iniconfig==1.1.1
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
selenium==4.4.2
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
wsproto==1.1.0
allure-pytest==2.9.45
allure-python-commons==2.9.45
requests~=2.28.1
nose~=1.3.7
bs4~=0.0.1
beautifulsoup4~=4.11.1
termcolor~=2.0.1